Refactor core logic for RL agent integration: - Remove unused `src/agent/rl/mod.rs`. - Add PyO3 support for Python bindings with new `lib.rs` entry point. - Update Cargo.toml and Cargo.lock with `pyo3` dependencies and dylib crate type. - Integrate RL-specific observation processing in `src/rl/obs.rs`. - Standardize combat mechanics: add cooldowns, tweak damage, and enhance safe zone logic. - Expand WorldConfig for fine-grained simulation parameters. - Migrate and refactor visualization modules under `viz::world`.

+++ src/item/pickup.rs
@@ -71,7 +71,7 @@ pub fn deposit_gold(
         if gold.is_empty() { continue; }
         if grid.get(pos.x, pos.y) != Some(crate::world::tile::Tile::Base(team.0)) { continue; }
 
-        let deposited = gold.0;
+        let deposited = gold.0 as u32;
         gold.0        = 0;
         score.0      += deposited;
         team_score.add(*team, deposited);
